esp32/PrimaryESP: Implement PositioningKalmanFilter (#115)
Implement a Kalman filter for a more accurate position estimation. The
filter takes raw position measurements from the positioning system, and
the estimated velocity of the vehicle based on motor actuation.

To make this work well, increase the loop frequency to 50 Hz and the CPU
frequency of the ESP32 to 240 MHz so that it can keep up under full
load.