clients/android: Fix a concurrency issue
Fix a bug that can cause an obstacle to have incorrect coordinates and
dimensions. This can happen while an obstacle is pending and another
obstacle is received. The client mistakes the newly received obstacle
for the pending one and assigns the wrong id to the pending obstacle.

The solution is to remove the pending obstacle and add the newly
received obstacle to the list of obstacles. There is still the problem
of removing the pending obstacle by mistake, but that is better than
having a misplaced obstacle.